% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areUnitsIdentical.R
\name{areUnitsIdentical}
\alias{areUnitsIdentical}
\title{Check whether units are identical following a specified list}
\usage{
areUnitsIdentical(vec1, vec2)
}
\arguments{
\item{vec1}{units to be checked against vec2, elementwise}

\item{vec2}{units to be checked against vec1, elementwise}
}
\value{
boolean
}
\description{
Check whether units are identical following a specified list
}
\author{
Oliver Richters
}
