% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFixUnits.R
\name{checkFixUnits}
\alias{checkFixUnits}
\title{Check units in IIASA submission by comparing mifdata to a project template}
\usage{
checkFixUnits(mifdata, template, logFile = NULL, failOnUnitMismatch = TRUE)
}
\arguments{
\item{mifdata}{quitte object or filename of mif file}

\item{template}{object provided by loadIIASAtemplate()}

\item{logFile}{filename of file for logging}

\item{failOnUnitMismatch}{boolean whether to fail in case of unit mismatches
recommended for submission, not used for generating templates}
}
\value{
quitte object with adapted mif data
}
\description{
Check units in IIASA submission by comparing mifdata to a project template
}
\author{
Oliver Richters
}
