% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingVars.R
\name{checkMissingVars}
\alias{checkMissingVars}
\title{Check whether all variables required by mapping are present in mifdata for given source}
\usage{
checkMissingVars(mifdata, mapping, sources = TRUE)
}
\arguments{
\item{mifdata}{data that can be read with as.quitte}

\item{mapping}{list of mapping that can be passed to getMappingVariables}

\item{sources}{source than can be passed to getMappingVariables}
}
\value{
an invisible vector with missing variables
}
\description{
Check whether all variables required by mapping are present in mifdata for given source
}
\author{
Oliver Richters
}
