% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNGFS.R
\name{checkNGFS}
\alias{checkNGFS}
\title{Check NGFS submission by comparing mif data to a template file (xlsx or yaml) provided by IIASA}
\usage{
checkNGFS(mifdata, iiasatemplate, logFile, generatePlots = TRUE)
}
\arguments{
\item{mifdata}{quitte object or filename of mif file}

\item{iiasatemplate}{filename of xlsx or yaml file provided by IIASA}

\item{logFile}{filename of file for logging. Set to NULL for stdout, set to FALSE for none.}

\item{generatePlots}{boolean whether to plot failing summations}
}
\value{
quitte object with adapted mif data
}
\description{
Check NGFS submission by comparing mif data to a template file (xlsx or yaml) provided by IIASA
}
\author{
Oliver Richters
}
