% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSummations.R
\name{checkSummations}
\alias{checkSummations}
\title{Checks for a run if the variables sum up as expected and logs spotted gaps}
\usage{
checkSummations(mifFile, outputDirectory = "output")
}
\arguments{
\item{mifFile}{path to the mif file to apply summation checks to}

\item{outputDirectory}{path to directory to place generated files (default: output)}
}
\description{
Checks for a run if the variables sum up as expected and logs spotted gaps
}
\author{
Falk Benke
}
