% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHistoricalData.R
\name{convertHistoricalData}
\alias{convertHistoricalData}
\title{Converts data in historical.mif to match project-specific variables and
regions so that it can be used for comparison in an intermodel comparios project}
\usage{
convertHistoricalData(mif, project, regionMapping = NULL)
}
\arguments{
\item{mif}{path to historical.mif to be converted}

\item{project}{name of the project, determines the mapping template to be loaded}

\item{regionMapping}{(optional) csv file with mapping of REMIND regions to
project regions, must contain two columns 'REMIND' and 'project_region'}
}
\description{
Converts data in historical.mif to match project-specific variables and
regions so that it can be used for comparison in an intermodel comparios project
}
\examples{
\dontrun{
data <- convertHistoricalData(
  mif = "path/to/histrical.mif",
  project = "NAVIGATE",
  regionMapping = "path/to/region_mapping_NAVIGATE.csv"
)
}
}
\author{
Falk Benke
}
