% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractReferenceYear.R
\name{extractReferenceYear}
\alias{extractReferenceYear}
\title{Extract reference year for price indices from unit}
\usage{
extractReferenceYear(unit)
}
\arguments{
\item{unit}{vector or string of units such as 'Index (2020 = 1)'}
}
\value{
vector or string of reference years such as '2020'
}
\description{
Extract reference year for price indices from unit
}
\author{
Oliver Richters
}
