% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMissing.R
\name{fillMissing}
\alias{fillMissing}
\title{Recursively calculate additional variables based on given summations}
\usage{
fillMissing(mifFile, summationsFile, iteration = 1)
}
\arguments{
\item{mifFile}{path to mif file}

\item{summationsFile}{in inst/summations folder that describes the required summation groups,
or path to summations file}

\item{iteration}{keeps track of number of recursive calls, leave to default}
}
\description{
Recursively calculate additional variables based on given summations
}
\author{
Falk Benke, Renato Rodrigues
}
