% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMissingSummations.R
\name{fillMissingSummations}
\alias{fillMissingSummations}
\title{Recursively calculate additional variables based on given summations and add
them to the given mif file}
\usage{
fillMissingSummations(mifFile, summationsFile, iteration = 1)
}
\arguments{
\item{mifFile}{path to mif file or a quitte object}

\item{summationsFile}{in inst/summations folder that describes the required summation groups,
or path to summations file}

\item{iteration}{keeps track of number of recursive calls, leave to default}
}
\description{
Recursively calculate additional variables based on given summations and add
them to the given mif file
}
\author{
Falk Benke, Renato Rodrigues
}
