% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillSummationPairs.R
\name{fillSummationPairs}
\alias{fillSummationPairs}
\title{add missing variable values if the value can be obtained from two other reported results.}
\usage{
fillSummationPairs(mifFile, summationsFile)
}
\arguments{
\item{mifFile}{path to mif file or a quitte object}

\item{summationsFile}{in inst/summations folder that describes the required summation groups,
or path to summations file}
}
\description{
add missing variable values if the value can be obtained from two other reported results.
}
\author{
Renato Rodrigues
}
