% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixOnRef.R
\name{fixOnRef}
\alias{fixOnRef}
\title{Checks for a run if it is correctly fixed on the reference run for t < startyear}
\usage{
fixOnRef(mif, mifRef, startyear, ret = "boolean", failfile = NULL)
}
\arguments{
\item{mif}{data or path to mif file of single scenario}

\item{mifRef}{data or path to mif file of reference scenario}

\item{startyear}{first time step for each mif and mifRef are expected to differ}

\item{ret}{"boolean", "fixed" or "fails", depending on what you want to get}

\item{failfile}{csv file to which failing check are written to}
}
\description{
Checks for a run if it is correctly fixed on the reference run for t < startyear
}
\author{
Oliver Richters
}
