% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceOutput.R
\name{generateIIASASubmission}
\alias{generateIIASASubmission}
\title{Generates an IIASA submission from REMIND runs by applying a project-specific mapping}
\usage{
generateIIASASubmission(
  mifDirectory,
  model,
  mapping,
  removeFromScen = NULL,
  addToScen = NULL,
  outputDirectory = "output",
  outputPrefix = "",
  logFile = "output/missing.log",
  generateSingleOutput = FALSE,
  outputFilename = "submission.mif"
)
}
\arguments{
\item{mifDirectory}{path to directory with mif files of a REMIND run}

\item{model}{name of model registered with IIASA}

\item{mapping}{path to mapping applied}

\item{removeFromScen}{string to be removed from scenario name (optional)}

\item{addToScen}{string to be added as prefix to scenario name (optional)}

\item{outputDirectory}{path to directory for the generated submission (default: output)}

\item{outputPrefix}{gets prepended to the file name of the generated file(s)}

\item{logFile}{path to the logfile with warnings (default: output/missing.log)}

\item{generateSingleOutput}{indicates whether the submission should be generated
into a single output file}

\item{outputFilename}{filename of the generated submission file,
if \code{generateSingleOutput} is set to TRUE (default: submission.mif)}
}
\description{
Generates an IIASA submission from REMIND runs by applying a project-specific mapping
}
\author{
Falk Benke
}
