% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMappingfile.R
\name{generateMappingfile}
\alias{generateMappingfile}
\title{Generate mapping from the template containing all the metadata}
\usage{
generateMappingfile(
  templates,
  fileName = NULL,
  remindVar,
  remindUnit,
  targetVar,
  targetUnit,
  factorCol,
  weightCol,
  spatialCol = "spatial",
  model = "REMIND-MAgPIE",
  commentFileName = NULL
)
}
\arguments{
\item{templates}{either a character or vector of templates to use as a starting point for creating the mapping}

\item{fileName}{name of the mapping file to be created, if provided, the file is created in the "output" folder,
otherwise a data frame is returned (optional)}

\item{remindVar}{column name in template containing the REMIND variable names}

\item{remindUnit}{column name in template containing the REMIND units}

\item{targetVar}{column name in template containing the target variable names}

\item{targetUnit}{column name in template containing the target units}

\item{factorCol}{column name in template containing conversion factors}

\item{weightCol}{column name in template containing weight factors}

\item{spatialCol}{column name in template containing regional restrictions
for reporting the corresponding variable (optional)}

\item{model}{exact name of the source model, used as column in comments file (optional)}

\item{commentFileName}{name of the comments file to be created, if provided,
the file is created in the "output" folder, otherwise a data frame is
returned (optional)}
}
\description{
Generate mapping from the template containing all the metadata
}
\author{
Falk Benke
}
