% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMapping.R
\name{getMapping}
\alias{getMapping}
\title{getMapping}
\usage{
getMapping(project = NULL)
}
\arguments{
\item{project}{name of requested mapping, or file name pointing to a mapping}
}
\description{
Retrieves latest mapping for a given project.
Mappings must contain the columns "variable", "unit", "piam_variable",
"piam_unit", "piam_factor".
Mappings are csv files with semicolon as a separator and no quotation marks
around fields, see main README.Rd file
}
\examples{
\dontrun{
getMapping("ECEMF")
getMapping("/path/to/mapping/file")
}
}
\author{
Falk Benke, Oliver Richters
}
