% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummations.R
\name{getSummations}
\alias{getSummations}
\title{Retrieves latest summation group file for a given project}
\usage{
getSummations(project = NULL, format = "dataframe")
}
\arguments{
\item{project}{name of the project of requested summation group, or summation group filename
which can be a csv file of a xlsx file where the first sheet containing a 'variable' column
is expected to have a 'components' column as well.
If project is a https://files.ece.iiasa.ac.at/*.xlsx  URL, it will be automatically downloaded.}

\item{format}{either "dataframe" or "list", the latter ignores the factor column}
}
\description{
Retrieves latest summation group file for a given project
}
\author{
Oliver Richters
}
