% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplateNames.R
\name{getTemplateNames}
\alias{getTemplateNames}
\title{Retrieves template file names}
\usage{
getTemplateNames(project = NULL)
}
\arguments{
\item{project}{name of the project of requested template, else will return all available templates.}
}
\value{
template file(s)
}
\description{
Retrieves template file names
}
\author{
Oliver Richters
}
