% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplateVariables.R
\name{getTemplateVariables}
\alias{getTemplateVariables}
\title{Retrieves all variables allocated to source potentially used
in mappings to project variables}
\usage{
getTemplateVariables(project, sources = TRUE)
}
\arguments{
\item{project}{name of the project of requested template}

\item{sources}{model abbreviation(s) used in 'source' column.
R = REMIND, M = MAgPIE, T = EDGE-T, B = Brick, C = Climate/MAGICC, TRUE = all}
}
\description{
Retrieves all variables allocated to source potentially used
in mappings to project variables
}
\examples{
getTemplateVariables("AR6", "RT")
}
\author{
Falk Benke, Oliver Richters
}
