% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingNames.R
\name{mappingNames}
\alias{mappingNames}
\title{Retrieves mapping file names}
\usage{
mappingNames(project = NULL)
}
\arguments{
\item{project}{name of the project of requested mapping. If not specified,
all existing mappings will be returned}
}
\value{
mapping file(s)
}
\description{
Retrieves mapping file names
}
\author{
Oliver Richters
}
