% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamInterfaces-package.R
\docType{package}
\name{piamInterfaces-package}
\alias{piamInterfaces}
\alias{piamInterfaces-package}
\title{piamInterfaces: Project specific interfaces to REMIND / MAgPIE}
\description{
Project specific interfaces to REMIND / MAgPIE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piam_interfaces}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

}
