% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntercomparison.R
\name{plotIntercomparison}
\alias{plotIntercomparison}
\title{Model intercomparison plots: area plots based on summation groups, line plots for further
variables. Creates a PDF for each model and scenario in the outputDirectory}
\usage{
plotIntercomparison(
  mifFile,
  outputDirectory = "output",
  summationsFile = "AR6",
  renameModels = NULL,
  lineplotVariables = TRUE,
  areaplotVariables = TRUE,
  interactive = FALSE,
  mainReg = "World",
  plotby = c("model", "scenario"),
  diffto = NULL,
  postfix = format(Sys.time(), "_\%Y-\%m-\%d_\%H.\%M.\%S")
)
}
\arguments{
\item{mifFile}{path to the mif or xlsx file to apply summation checks to, or quitte object}

\item{outputDirectory}{path to directory to place one PDF for each model and scenario}

\item{summationsFile}{in inst/summations folder that describes the required summation groups}

\item{renameModels}{vector with oldname = newname}

\item{lineplotVariables}{vector with variable names for lineplots or filenames
of files containing a 'variable' column (or both)}

\item{areaplotVariables}{vector with variable names for areaplot or filenames
of files containing a 'variable' column. Only those available in the summationsFile can be plotted.}

\item{interactive}{boolean whether you want to select variables, regions and models to be plotted}

\item{mainReg}{region name of main region to be passed to mip}

\item{plotby}{whether you would like to have everything plotted by scenario, model and/or onefile.
set to NULL to be asked.}

\item{diffto}{if specified, the difference to this scenario is calculated and plotted}

\item{postfix}{to the filename, defaults to something like "_2024-09-05_12.47.28"}
}
\description{
Model intercomparison plots: area plots based on summation groups, line plots for further
variables. Creates a PDF for each model and scenario in the outputDirectory
}
\examples{

\dontrun{
plotIntercomparison(quitte::quitte_example_dataAR6,
                    lineplotVariables = c("Temperature|Global Mean", "Population"))
}

}
\author{
Oliver Richters
}
