% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceIndicesAdd.R
\name{priceIndicesAdd}
\alias{priceIndicesAdd}
\title{Add price index}
\usage{
priceIndicesAdd(mifdata, priceIndices, scenBase = NULL, referenceYear = 2020)
}
\arguments{
\item{mifdata}{file or data that can be converted into quitte object}

\item{priceIndices}{vector of missing price index variable names}

\item{scenBase}{optional scenario name of baseline scenario used to calculate index}

\item{referenceYear}{in which index = 1}
}
\description{
Add price index
}
\author{
Oliver Richters
}
