% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceIndicesFix.R
\name{priceIndicesFix}
\alias{priceIndicesFix}
\title{Fixes price indices with wrong reference year}
\usage{
priceIndicesFix(mifdata, priceIndices, referenceYear = 2020)
}
\arguments{
\item{mifdata}{file or data that can be converted into quitte object}

\item{priceIndices}{vector of missing price index variable names}

\item{referenceYear}{in which index = 1}
}
\description{
Fixes price indices with wrong reference year
}
\author{
Oliver Richters
}
