% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceIndicesIIASA.R
\name{priceIndicesIIASA}
\alias{priceIndicesIIASA}
\title{Add Price|\emph{|Index variables requested in iiasatemplate but missing in data,
if Price|} is present in data. Extracts reference year automatically from unit}
\usage{
priceIndicesIIASA(mifdata, iiasatemplate, scenBase = NULL)
}
\arguments{
\item{mifdata}{file or data that can be converted into quitte object}

\item{iiasatemplate}{filename of xlsx or yaml file provided by IIASA}

\item{scenBase}{optional scenario name of baseline scenario used to calculate index}
}
\description{
Add Price|\emph{|Index variables requested in iiasatemplate but missing in data,
if Price|} is present in data. Extracts reference year automatically from unit
}
\author{
Oliver Richters
}
