% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMifs.R
\name{readMifs}
\alias{readMifs}
\title{Pass a character vector containing filenames and directories.
Returns data from all files and all '.mif' files in the directories.}
\usage{
readMifs(...)
}
\arguments{
\item{...}{path to mif files or directories with mif files of a REMIND run or quitte object}
}
\description{
Pass a character vector containing filenames and directories.
Returns data from all files and all '.mif' files in the directories.
}
\examples{
\dontrun{
# Simple use. Generates submission file in output folder:
readMifs(
  mifs = "/path/to/REMIMD/mifs",
)
}
}
\author{
Falk Benke, Oliver Richters
}
