% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePlus.R
\name{removePlus}
\alias{removePlus}
\title{Remove |+|, |++| etc. from variable names}
\usage{
removePlus(x)
}
\arguments{
\item{x}{vector with variable names}
}
\value{
varname without any plusses human-readable summary to the user
}
\description{
Remove |+|, |++| etc. from variable names
}
\examples{
#' removePlus(c("FE|+|CDR", "FE|CDR|DACCS"))
}
\author{
Oliver Richters
}
