% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameOldInMappings.R
\name{renameOldInMappings}
\alias{renameOldInMappings}
\title{renameOldInMappings}
\usage{
renameOldInMappings(folder = ".")
}
\arguments{
\item{folder}{piamInterfaces folder}
}
\description{
Loads "renamed_piam_variables.csv" and replaces all instances of old_name
in the piam_variable columns of all mapping files
Handle with care, better commit your changes before using it
Best run this in your piamInterfaces folder:
Rscript -e 'devtools::load_all(); renameOldInMappings()'
}
\author{
Oliver Richters
}
