% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLogFile.R
\name{setLogFile}
\alias{setLogFile}
\title{Generate valid path to logFile and make sure the outputDirectory exists.
If logFile is just a file name without any further path info, put logFile in outputDirectory}
\usage{
setLogFile(outputDirectory = NULL, logFile = NULL)
}
\arguments{
\item{outputDirectory}{path to directory to place generated files}

\item{logFile}{path or name for log file}
}
\description{
Generate valid path to logFile and make sure the outputDirectory exists.
If logFile is just a file name without any further path info, put logFile in outputDirectory
}
\author{
Oliver Richters
}
