% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summationsNames.R
\name{summationsNames}
\alias{summationsNames}
\title{Retrieves summation group file names}
\usage{
summationsNames(project = NULL)
}
\arguments{
\item{project}{name of the project of requested summation file. If not specified,
all existing summation files will be returned}
}
\value{
template file(s)
}
\description{
Retrieves summation group file names
}
\author{
Oliver Richters
}
