% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateNames.R
\name{templateNames}
\alias{templateNames}
\title{Retrieves template file names}
\usage{
templateNames(project = NULL)
}
\arguments{
\item{project}{name of the project of requested template. If not specified,
all existing templates will be returned}
}
\value{
template file(s)
}
\description{
Retrieves template file names
}
\author{
Oliver Richters
}
