% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePlus.R
\name{updatePlusUnit}
\alias{updatePlusUnit}
\title{updatePlusUnit}
\usage{
updatePlusUnit(mif, folder = ".")
}
\arguments{
\item{mif}{data file or quitte object}

\item{folder}{piamInterfaces folder}
}
\description{
Loads data and replaces all variables with the |+| notation used in this data.
Pick a mif file or quitte object and run:
Rscript -e 'devtools::load_all(); updatePlusUnit("your.mif")'
}
\author{
Oliver Richters
}
