% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMapping.R
\name{writeMapping}
\alias{writeMapping}
\title{writeMapping}
\usage{
writeMapping(mapping, filename)
}
\arguments{
\item{mapping}{mapping data.frame}

\item{filename}{filename.}
}
\description{
write data.frame obtained from getMapping() to mapping file.
}
\author{
Oliver Richters
}
