% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areUnitsIdentical.R
\name{areUnitsIdentical}
\alias{areUnitsIdentical}
\title{Check whether units are identical following a specified list}
\usage{
areUnitsIdentical(...)
}
\arguments{
\item{...}{units to be checked}
}
\value{
boolean
}
\description{
Check whether units are identical following a specified list
}
\author{
Oliver Richters
}
