% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIIASASubmission.R
\name{checkIIASASubmission}
\alias{checkIIASASubmission}
\title{Check IIASA submission by comparing mif data to a template file (xlsx or yaml) provided by IIASA}
\usage{
checkIIASASubmission(
  mifdata,
  iiasatemplate,
  logFile = NULL,
  failOnUnitMismatch = TRUE
)
}
\arguments{
\item{mifdata}{quitte object or filename of mif file}

\item{iiasatemplate}{filename of xlsx or yaml file provided by IIASA}

\item{logFile}{filename of file for logging}

\item{failOnUnitMismatch}{boolean whether to fail in case of unit mismatches
recommended for submission, not used for generating templates}
}
\value{
quitte object with adapted mif data
}
\description{
Check IIASA submission by comparing mif data to a template file (xlsx or yaml) provided by IIASA
}
\examples{
\dontrun{
# Simple use. Generates submission file in output folder:
checkIIASASubmission(
  mifdata = "file.mif",
  iiasatemplate = "template.xlsx",
  logFile = "logFile.txt"
)
}
}
\author{
Oliver Richters
}
