% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingVars.R
\name{checkMissingVars}
\alias{checkMissingVars}
\title{Check whether all variables required by mapping are present in mifdata for given source}
\usage{
checkMissingVars(mifdata, mapping = TRUE, sources = TRUE)
}
\arguments{
\item{mifdata}{data that can be read with as.quitte}

\item{mapping}{name of the project of requested mappings
such as c('AR6', 'AR6_NGFS') or 'mapping.csv'. Use TRUE for all mappings.}

\item{sources}{model abbreviation(s) used in 'source' column.
R = REMIND, M = MAgPIE, T = EDGE-T, B = Brick, C = Climate/MAGICC, TRUE = all}
}
\value{
an invisible vector with missing variables
}
\description{
Check whether all variables required by mapping are present in mifdata for given source
}
\author{
Oliver Richters
}
