% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSummations.R
\name{checkSummations}
\alias{checkSummations}
\title{Checks for a run if the variables sum up as expected and logs spotted gaps}
\usage{
checkSummations(
  mifFile,
  outputDirectory = ".",
  template = NULL,
  summationsFile = NULL,
  logFile = NULL,
  logAppend = FALSE,
  generatePlots = FALSE,
  mainReg = "World",
  dataDumpFile = "checkSummations.csv",
  plotprefix = NULL,
  absDiff = 0.001,
  relDiff = 1,
  roundDiff = TRUE,
  csvSeparator = ";"
)
}
\arguments{
\item{mifFile}{path to the mif file to apply summation checks to, or quitte object}

\item{outputDirectory}{path to directory to place logFile and dataDumpFile.}

\item{template}{mapping to be loaded, used to print the piam_variable corresponding to the data variables}

\item{summationsFile}{in inst/summations folder that describes the required summation groups
if set to 'extractVariableGroups', tries to extract summations from variables with + notation}

\item{logFile}{file where human-readable summary is saved. If NULL, write to stdout. If FALSE, don't log.}

\item{logAppend}{boolean whether to append or overwrite logFile}

\item{generatePlots}{boolean whether pdfs to compare data are generated. Requires outputDirectory.}

\item{mainReg}{main region for the plot generation}

\item{dataDumpFile}{file where data.frame with the data analysis is saved. Requires outputDirectory.
If NULL, result is returned.}

\item{plotprefix}{added before filename}

\item{absDiff}{threshold for absolute difference between parent variable and summation}

\item{relDiff}{threshold (in percent) for relative difference between parent variable and summation}

\item{roundDiff}{should the absolute and relative differences in human-readable summary and dataDumpFile
be rounded? The returned object always contains unrounded values.}

\item{csvSeparator}{separator for dataDumpFile, defaults to semicolon}
}
\description{
Checks for a run if the variables sum up as expected and logs spotted gaps
}
\author{
Falk Benke, Oliver Richters
}
