% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUnitFactor.R
\name{checkUnitFactor}
\alias{checkUnitFactor}
\title{Check unit factor in template}
\usage{
checkUnitFactor(template, logFile = NULL, failOnUnitMismatch = TRUE)
}
\arguments{
\item{template}{object provided by loadIIASAtemplate()}

\item{logFile}{filename of file for logging}

\item{failOnUnitMismatch}{boolean whether to fail in case of unit mismatches
recommended for submission, not used for checking mapping}
}
\value{
quitte object with adapted mif data
}
\description{
Check unit factor in template
}
\author{
Oliver Richters
}
