% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHistoricalData.R
\name{convertHistoricalData}
\alias{convertHistoricalData}
\title{Converts data in historical.mif to match project-specific variables and
regions so that it can be used for comparison in an intermodel comparison project}
\usage{
convertHistoricalData(mif, project, regionMapping = NULL)
}
\arguments{
\item{mif}{quitte object with historical data or path to historical.mif}

\item{project}{name of the project, determines the mapping to be loaded}

\item{regionMapping}{(optional) csv file with mapping of REMIND regions or ISO to
project regions, must contain two columns:
One can be called 'REMIND' or 'CountryCode' and contains the regions in the data.
The second can be called 'project_region' or 'RegionCode', to which the first is mapped.
You can also specify the filename part before the .csv from inst/regionmapping}
}
\description{
Converts data in historical.mif to match project-specific variables and
regions so that it can be used for comparison in an intermodel comparison project
}
\examples{
\dontrun{
data <- convertHistoricalData(
  mif = "path/to/historical.mif",
  project = "NAVIGATE",
  regionMapping = "path/to/region_mapping_NAVIGATE.csv"
)
}
}
\author{
Falk Benke
}
