% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIIASASubmission.R
\name{generateIIASASubmission}
\alias{generateIIASASubmission}
\title{Generates an IIASA submission from REMIND or MAgPIE runs by applying a project-specific mapping.
The script starts from 'mifs' which can be a directory with mif files, a vector of files or a
quitte object. In outputDirectory/outputFilename, you will get the data in a joint xlsx or mif file.}
\usage{
generateIIASASubmission(
  mifs = ".",
  mapping = NULL,
  model = NULL,
  mappingFile = NULL,
  removeFromScen = NULL,
  addToScen = NULL,
  outputDirectory = "output",
  logFile = paste0(gsub("\\\\.[a-zA-Z]+$", "_log.txt", outputFilename)),
  outputFilename = "submission.xlsx",
  iiasatemplate = NULL,
  generatePlots = FALSE,
  timesteps = c(seq(2005, 2060, 5), seq(2070, 2100, 10)),
  generateSingleOutput = TRUE
)
}
\arguments{
\item{mifs}{path to mif files or directories with mif files of a REMIND run,
or quitte object}

\item{mapping}{mapping template names such as c("AR6", "AR6_NGFS") or a mapping template filename.
If NULL and no existing mappingFile provided, user is asked.
If specified, the parameter mappingFile is overwritten.}

\item{model}{name of model as registered with IIASA}

\item{mappingFile}{path to a mapping file generated by generateMappings(). Requires mapping = NULL!}

\item{removeFromScen}{regular expression to be removed from scenario name (optional). Example: '_d50|d95'}

\item{addToScen}{string to be added as prefix to scenario name (optional)}

\item{outputDirectory}{path to directory for the generated submission (default: output)}

\item{logFile}{path to the logfile with warnings as passed to generateMappingfile, checkIIASASubmission
(default: outputDirectory/submission_missing.log). Set to FALSE for none.}

\item{outputFilename}{filename of the generated submission. Must be mif or xlsx file.}

\item{iiasatemplate}{optional filename of xlsx or yaml file provided by IIASA
used to delete superfluous variables and adapt units}

\item{generatePlots}{boolean, whether to generate plots of failing summation checks}

\item{timesteps}{timesteps that are accepted in final submission}

\item{generateSingleOutput}{has no effect and is only kept for backwards-compatibility}
}
\description{
To provide the mapping, two options exist:
\itemize{
\item If you already have a mapping as generated by generateMappings(), set mapping = NULL and
mappingFile to the desired filename.
\item If you want to generate the mapping from one or more mapping templates from the inst/templates folder,
set mapping = c("AR6", "AR6_NGFS") or so. mappingFile is ignored then.
}
}
\details{
iiasatemplate is a xlsx or yaml file provided by IIASA with the variable + unit definitions that are
accepted in the database. The function 'priceIndicesIIASA' will be called to calculate price indices
that are missing or with the wrong base year. 'checkIIASASubmission' will be called to remove all variables
that are not accepted in the database.

For all elements of the parameter mapping that contain a summation file in inst/summations,
the function 'checkSummations' is called to verify variable summation checks.

To alter the data, you can use those parameters: model, addToScen, removeFromScen and timesteps.

For a broader overview of the submission process, consult
https://github.com/remindmodel/remind/blob/develop/tutorials/13_Submit_to_IIASA_database.md
}
\examples{
\dontrun{
# Simple use. Generates submission file in output folder:
generateIIASASubmission(
  mifs = "/path/to/REMIMD/mifs",
  model = "REMIND-MAgPIE 2.1-4.2",
  mappingFile = "output/template_navigate.csv"
)
}
}
\author{
Falk Benke, Oliver Richters
}
