% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIIASASubmission.R
\name{generateIIASASubmission}
\alias{generateIIASASubmission}
\title{Generates an IIASA submission from REMIND runs by applying a project-specific mapping}
\usage{
generateIIASASubmission(
  mifs = ".",
  mapping = NULL,
  model = "REMIND 3.0",
  mappingFile = NULL,
  removeFromScen = NULL,
  addToScen = NULL,
  outputDirectory = "output",
  outputPrefix = "",
  logFile = "output/missing.log",
  generateSingleOutput = TRUE,
  outputFilename = "submission.mif",
  iiasatemplate = NULL
)
}
\arguments{
\item{mifs}{path to mif files or directories with mif files of a REMIND run}

\item{mapping}{mapping template names such as c("AR6", "AR6_NGFS"). If NULL, user is asked}

\item{model}{name of model registered with IIASA}

\item{mappingFile}{path to mapping. If NULL, mapping is generated based on param mapping}

\item{removeFromScen}{string to be removed from scenario name (optional)}

\item{addToScen}{string to be added as prefix to scenario name (optional)}

\item{outputDirectory}{path to directory for the generated submission (default: output)}

\item{outputPrefix}{gets prepended to the file name of the generated file(s)}

\item{logFile}{path to the logfile with warnings (default: output/missing.log)}

\item{generateSingleOutput}{indicates whether the submission should be generated
into a single output file}

\item{outputFilename}{filename of the generated submission file,
if \code{generateSingleOutput} is set to TRUE (default: submission.mif)}

\item{iiasatemplate}{optional filename of xlsx or yaml file provided by IIASA
used to delete superfluous variables and adapt units}
}
\description{
Generates an IIASA submission from REMIND runs by applying a project-specific mapping
}
\examples{
\dontrun{
# Simple use. Generates submission file in output folder:
generateIIASASubmission(
  mifs = "/path/to/REMIMD/mifs",
  model = "REMIND-MAgPIE 2.1-4.2",
  mappingFile = "output/template_navigate.csv",
  generateSingleOutput = TRUE
)
}
}
\author{
Falk Benke, Oliver Richters
}
