% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIIASASubmission.R
\name{generateIIASASubmission}
\alias{generateIIASASubmission}
\title{Generates an IIASA submission from REMIND runs by applying a project-specific mapping}
\usage{
generateIIASASubmission(
  mifs = ".",
  mapping = NULL,
  model = "REMIND 3.0",
  mappingFile = NULL,
  removeFromScen = NULL,
  addToScen = NULL,
  outputDirectory = "output",
  logFile = "output/missing.log",
  outputFilename = "submission.xlsx",
  iiasatemplate = NULL,
  generatePlots = FALSE,
  timesteps = c(seq(2005, 2060, 5), seq(2070, 2100, 10)),
  generateSingleOutput = TRUE
)
}
\arguments{
\item{mifs}{path to mif files or directories with mif files of a REMIND run}

\item{mapping}{mapping template names such as c("AR6", "AR6_NGFS"). If NULL, user is asked}

\item{model}{name of model registered with IIASA}

\item{mappingFile}{path to mapping. If NULL, mapping is generated based on param mapping}

\item{removeFromScen}{string to be removed from scenario name (optional)}

\item{addToScen}{string to be added as prefix to scenario name (optional)}

\item{outputDirectory}{path to directory for the generated submission (default: output)}

\item{logFile}{path to the logfile with warnings (default: output/missing.log)}

\item{outputFilename}{filename of the generated submission mif or xlsx file.}

\item{iiasatemplate}{optional filename of xlsx or yaml file provided by IIASA
used to delete superfluous variables and adapt units}

\item{generatePlots}{boolean, whether to generate plots of failing summation checks}

\item{timesteps}{timesteps that are accepted in final submission}

\item{generateSingleOutput}{has no effect and is only kept for backwards-compatibility}
}
\description{
Generates an IIASA submission from REMIND runs by applying a project-specific mapping
}
\examples{
\dontrun{
# Simple use. Generates submission file in output folder:
generateIIASASubmission(
  mifs = "/path/to/REMIMD/mifs",
  model = "REMIND-MAgPIE 2.1-4.2",
  mappingFile = "output/template_navigate.csv"
)
}
}
\author{
Falk Benke, Oliver Richters
}
