% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMappingfile.R
\name{generateMappingfile}
\alias{generateMappingfile}
\title{Generate mapping from the template containing all the metadata}
\usage{
generateMappingfile(
  templates = NULL,
  outputDirectory = "output",
  fileName = NULL,
  commentFileName = NULL,
  targetVar = "Variable",
  targetUnit = "Unit",
  remindVar = "r30m44",
  remindUnit = "r30m44_unit",
  factorCol = "r30m44_factor",
  weightCol = NULL,
  spatialCol = "spatial",
  model = "REMIND-MAgPIE"
)
}
\arguments{
\item{templates}{either a character or vector of project templates to use as a starting point for
creating the mapping, valid projects are "AR6", "SHAPE", "NAVIGATE"}

\item{outputDirectory}{path to directory to place generated files (default: output)}

\item{fileName}{name of the mapping file to be created, if provided, the file is created in the "output" folder,
otherwise a data frame is returned (optional)}

\item{commentFileName}{name of the comments file to be created, if provided,
the file is created in the "output" folder, otherwise a data frame is
returned (optional)}

\item{targetVar}{column name in template containing the target variable names}

\item{targetUnit}{column name in template containing the target units (default: Unit)}

\item{remindVar}{column name in template containing the REMIND variable names (default: r30m44)}

\item{remindUnit}{column name in template containing the REMIND units (default: r30m44_unit)}

\item{factorCol}{column name in template containing conversion factors (default: r30m44_factor)}

\item{weightCol}{column name in template containing weight factors (optional)}

\item{spatialCol}{column name in template containing regional restrictions
for reporting the corresponding variable (optional)}

\item{model}{exact name of the source model, used as column in comments file (default: REMIND-MAgPIE)}
}
\description{
Generate mapping from the template containing all the metadata
}
\examples{
\dontrun{
# Simple use. Creates NAVIGATE mapping and saves it to /output/template_navigate.csv:
generateMappingfile(
  templates = "NAVIGATE",
  targetVar = "Variable_NAVIGATE",
  fileName = "template_navigate.csv"
)
# More complex use. Creates combined mapping from NAVIGATE and SHAPE template:
generateMappingfile(
  templates = c("NAVIGATE", "SHAPE"),
  targetVar = c("Variable_NAVIGATE", "Variable_SHAPE"),
  fileName = "template_combined.csv"
)
}
}
\author{
Falk Benke
}
