% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getREMINDTemplateVariables.R
\name{getREMINDTemplateVariables}
\alias{getREMINDTemplateVariables}
\title{Retrieves all REMIND variables potentially used in mappings to project variables}
\usage{
getREMINDTemplateVariables(project, excludeMAGICC = TRUE)
}
\arguments{
\item{project}{name of the project of requested template}

\item{excludeMAGICC}{exclude MAGICC variables used in template}
}
\description{
Retrieves all REMIND variables potentially used in mappings to project variables
}
\author{
Falk Benke
}
