% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMappingVariables.R
\name{getREMINDTemplateVariables}
\alias{getREMINDTemplateVariables}
\title{legacy function to be used by remind2}
\usage{
getREMINDTemplateVariables(project)
}
\arguments{
\item{project}{name of the project of requested mapping}
}
\description{
legacy function to be used by remind2
}
