% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummations.R
\name{getSummations}
\alias{getSummations}
\title{Retrieves latest summation group file for a given project}
\usage{
getSummations(project = NULL)
}
\arguments{
\item{project}{name of the project of requested summation group, or summation group filename}
}
\description{
Retrieves latest summation group file for a given project
}
\author{
Oliver Richters
}
