% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplate.R
\name{getTemplate}
\alias{getTemplate}
\title{Retrieves latest template for a given project}
\usage{
getTemplate(project)
}
\arguments{
\item{project}{name of the project of requested template}
}
\description{
Retrieves latest template for a given project
}
\author{
Falk Benke
}
