% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMapping.R
\name{getTemplate}
\alias{getTemplate}
\title{for backwards compatibility}
\usage{
getTemplate(project = NULL)
}
\arguments{
\item{project}{name of requested mapping, or file name pointing to a mapping}
}
\description{
for backwards compatibility
}
