% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplate.R
\name{getTemplate}
\alias{getTemplate}
\title{getTemplate}
\usage{
getTemplate(project = NULL)
}
\arguments{
\item{project}{name of requested template, or file name pointing to a template}
}
\description{
Retrieves latest template for a given project.
Templates must contain the columns "Variable", "Unit", "piam_variable",
"piam_unit", "piam_factor".
Templates are csv files with semicolon as a separator and no quotation marks
around fields.
}
\author{
Falk Benke, Oliver Richters
}
