% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceround.R
\name{niceround}
\alias{niceround}
\title{niceround returns human-readable string for big and small numbers
Show everything before the dot and at least 'digits' significant digits}
\usage{
niceround(x, digits = 3)
}
\arguments{
\item{x}{some number}

\item{digits}{number of significant digits that are definitely kept}
}
\description{
niceround returns human-readable string for big and small numbers
Show everything before the dot and at least 'digits' significant digits
}
