% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceround.R
\name{niceround}
\alias{niceround}
\title{niceround returns human-readable string for big and small numbers
If abs(x) > 1, show everything before the dot plus 2 digits after. For abs(x) < 1, show 3 digits}
\usage{
niceround(x)
}
\arguments{
\item{x}{some number}
}
\description{
niceround returns human-readable string for big and small numbers
If abs(x) > 1, show everything before the dot plus 2 digits after. For abs(x) < 1, show 3 digits
}
