% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntercomparison.R
\name{plotIntercomparison}
\alias{plotIntercomparison}
\title{Model intercomparison plot}
\usage{
plotIntercomparison(
  mifFile,
  outputDirectory = ".",
  summationsFile = "AR6",
  renameModels = NULL,
  lineplotVariables = c("Temperature|Global Mean")
)
}
\arguments{
\item{mifFile}{path to the mif or xlsx file to apply summation checks to, or quitte object}

\item{outputDirectory}{path to directory to place logFile and dataDumpFile}

\item{summationsFile}{in inst/summations folder that describes the required summation groups}

\item{renameModels}{vector with oldname = newname}

\item{lineplotVariables}{vector with variable names for additional lineplots}
}
\description{
Model intercomparison plot
}
\examples{

\dontrun{
plotIntercomparison(quitte::quitte_example_dataAR6,
                    lineplotVariables = c("Temperature|Global Mean", "Population"))
}

}
\author{
Oliver Richters
}
