% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableInfo.R
\name{variableInfo}
\alias{variableInfo}
\title{Provide information on variable, its mappings and summation groups}
\usage{
variableInfo(
  varname,
  mif = NULL,
  template = NULL,
  remindVar = "piam_variable",
  remindUnit = "piam_unit"
)
}
\arguments{
\item{varname}{string with variable name}

\item{mif}{filename of miffile}

\item{template}{template shortcut (AR6, NAVIGATE). NULL means all}

\item{remindVar}{column name of variable in templates (default: piam_variable)}

\item{remindUnit}{column name of unit in templates (default: piam_unit)}
}
\value{
prints human-readable summary to the user
}
\description{
Provide information on variable, its mappings and summation groups
}
\examples{
# Simple use. prints human-readable summary to the reader on Emi|CO2:
variableInfo(
  "Emi|CO2"
)
}
\author{
Oliver Richters
}
